Moved all types to types.ts
